Homework 12 is corrected
