+++ src/homework12/Author.java
@@ -1,0 +1,17 @@
+package homework12;
+
+public class Author {
+    private String firstName;
+    private String lastName;
+    public Author(String firstName, String lastName){
+        this.firstName = firstName;
+        this.lastName = lastName;
+    }
+    public String getFirstName(){
+        return this.firstName;
+    }
+
+    public String getLastName(){
+        return this.lastName;
+    }
+}
